Add the Scathlete library to project
created local library for scala access to to strava api. going to leave this as a local dependency for now. required updates to the scala version bringing routescoop-api up to 2.12.12. added some builder methods to the api models to convert from strava types. need to figure out how to test the scathlete addition given that it requires a recent strava access token for a bonafide user.