Support activity stats updates in api
In order to make changes to the activity stats manually, a PUT endpoint is necessary for activity stats. This will enable a client application to send a payload to alter the stats generated for an activity in case one of the metrics cannot be generated. For example, if no power data, we need to manually assign the training load for the activity.